Add vagrant environment
Vagrant environment sets up OpenShift environment for development.
The liveliness and readiness probe for koji is still not working.
This causes error to be thrown at the end of the mbbox script.

Koji still needs some work, but this environment should be enough to
start working with this script locally.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>